refactor(utils) - rename server utils
