+++ internal/upstreamserver/handlers.go
@@ -5,15 +5,15 @@ import (
 	"net/http"
 
 	"github.com/ronfelsenfeld/go-proxy/internal/logger"
-	apiUtils "github.com/ronfelsenfeld/go-proxy/internal/utils/apiutils"
+	serverUtils "github.com/ronfelsenfeld/go-proxy/internal/utils/serverutils"
 )
 
 func testHandler(responseWriter http.ResponseWriter, request *http.Request) {
 	logger.Info.Printf("📥 Received %s request at %s", request.Method, request.URL.Path)
 
-	apiUtils.PrintRequestHeaders(request)
+	serverUtils.PrintRequestHeaders(request)
 
-	requestBody, err := apiUtils.DecodeRequestBody(request)
+	requestBody, err := serverUtils.DecodeRequestBody(request)
 	if err != nil {
 		logger.Error.Printf("❌ %s", err.Error())
 		http.Error(responseWriter, err.Error(), http.StatusBadRequest)
